AIM-74 refactor: image, video module code refactor(function separation)
