+++ vlm_model_gpt4o-mini/utils/analysis.py
@@ -7,7 +7,7 @@ import numpy as np
 import math
 from config import SYSTEM_INSTRUCTION
 from constants.behaviors import PROBLEMATIC_BEHAVIORS
-from utils.image_utils import encode_image
+from utils.encoding_image import encode_image
 
 # OpenAI 모듈을 client로 정의
 client = openai
