+++ vlm_model_gpt4o-mini/utils/encoding_image.py
@@ -1,4 +1,4 @@
-# utils/image_utils.py
+# utils/encoding_iamge.py
 
 import cv2
 import base64
@@ -33,9 +33,3 @@ def encode_image(image: np.ndarray, max_size: tuple = (256, 256), quality: int =
     except Exception as e:
         print(f"이미지 인코딩 중 오류 발생: {e}")
         return None
-
-def generate_user_prompt(img_type: str, img_b64_str: str) -> str:
-    """
-    주어진 이미지 데이터와 타입을 기반으로 비언어적 행동 평가를 요청하는 사용자 프롬프트를 생성합니다.
-    """
-    return f"Please evaluate the presenter's non-verbal behavior.\n\nImage data: data:{img_type};base64,{img_b64_str}"
